Cleanup
+ Use `var` for all local variables excluding primitives and arrays
+ Use method references instead of lambdas where possible
+ Convert to enhanced `for` loops where possible
+ Convert fields into local variables if the use is only local
+ Add missing `@Override` annotations to implementations of interface methods
+ Remove unnecessary casts
+ Remove redundant modifiers
+ Remove variable assignments before return where possible
+ Correct indentation
+ Remove redundant super() call in constructor
+ Use pattern matching for instanceof
+ Use lambdas where possible
+ Use diamond operator where possible

+++ core/src/hitonoriol/madsand/util/TimeUtils.java
@@ -21,15 +21,15 @@ public class TimeUtils {
 	public static Timer.Task scheduleTask(Runnable task, float delaySec) {
 		return scheduleTask(Timer.instance(), task, delaySec);
 	}
-	
+
 	public static Timer.Task scheduleTask(Timer timer, Timer.Task task, float delaySec) {
 		return timer.scheduleTask(task, delaySec);
 	}
-	
+
 	public static Timer.Task scheduleTask(Timer.Task task, float delaySec) {
 		return scheduleTask(Timer.instance(), task, delaySec);
 	}
-	
+
 	public static Timer.Task scheduleTask(Runnable task) {
 		return scheduleTask(task, Gui.DELAY);
 	}
@@ -43,6 +43,6 @@ public class TimeUtils {
 	}
 
 	public static float timeToExecution(Timer.Task task) {
-		return (float) (task.getExecuteTimeMillis() - System.nanoTime() / 1000000) / 1000f;
+		return (task.getExecuteTimeMillis() - System.nanoTime() / 1000000) / 1000f;
 	}
 }
